requirements.txt
# SpecRepro requirements
# Core LLM clients
anthropic>=0.40.0
openai>=1.30.0

# PDF parsing (at least one of these needed for .pdf papers)
pdfminer.six>=20221105

# Deep learning (for generated code)
torch>=2.0.0
torchvision>=0.15.0

# Data / utilities
numpy>=1.24.0
